requirements.txt
python 3.8
python-decouple
pandas
psycopg2
sqlalchemy
